9th commit - Exercise 2 to send

+++ src/main/java/com/przemek/patronage/Equipment/EquipmentService.java
@@ -1,30 +1,46 @@
 package com.przemek.patronage.Equipment;
 
+import com.przemek.patronage.ConferenceRoom.ConferenceRoomRepository;
+import com.przemek.patronage.Exceptions.NoSuchIdException;
+import lombok.var;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 @Service
 public class EquipmentService {
     private EquipmentRepository equipmentRepository;
+    private ConferenceRoomRepository conferenceRooms;
+
+    public EquipmentService() {
+
+    }
 
     @Autowired
-    public EquipmentService(EquipmentRepository equipmentRepository) {
+    public EquipmentService(EquipmentRepository equipmentRepository, ConferenceRoomRepository conferenceRooms) {
         this.equipmentRepository = Objects.requireNonNull(equipmentRepository, "must be defined.");
+        this.conferenceRooms = Objects.requireNonNull(conferenceRooms, "must be defined.");
     }
 
     public List<Equipment> findAll() {
         return equipmentRepository.findAll();
     }
 
-    public void save(Equipment newEquipment) {
+    public void save(Equipment newEquipment, Long id) throws NoSuchIdException {
+        if (conferenceRooms.findById(id).equals(Optional.empty())) {
+            throw new NoSuchIdException("The Conference room with id given doesn't exist in the base.");
+        }
         if (!newEquipment.isPhone()) {
             newEquipment.setInternalNumber(null);
             newEquipment.setExternalNumber(null);
             newEquipment.setConnections(null);
         }
+        var room = conferenceRooms.findById(id).get();
+        room.setEquipment(newEquipment);
+        newEquipment.setConferenceroom(room);
         equipmentRepository.save(newEquipment);
     }
 
@@ -34,11 +50,10 @@ public class EquipmentService {
                 .map(equipment -> {
                     equipment.setProjectorName(newEquipment.getProjectorName());
                     equipment.setPhone(newEquipment.isPhone());
-                    if (equipment.isPhone()) {
-                        equipment.setInternalNumber(newEquipment.getInternalNumber());
-                        equipment.setExternalNumber(newEquipment.getExternalNumber());
-                        equipment.setConnections(newEquipment.getConnections());
-                    }
+                    equipment.setInternalNumber(newEquipment.getInternalNumber());
+                    equipment.setExternalNumber(newEquipment.getExternalNumber());
+                    equipment.setConnections(newEquipment.getConnections());
+                    equipment.setConferenceroom(newEquipment.getConferenceroom());
                     return equipmentRepository.save(equipment);
                 })
                 .orElseGet(() -> {
@@ -47,7 +62,10 @@ public class EquipmentService {
                 });
     }
 
-    public void delete(Long id) {
-        equipmentRepository.deleteById(id);
+    public void delete(Long id) throws NoSuchIdException {
+        if (equipmentRepository.findById(id).equals(Optional.empty())) {
+            throw new NoSuchIdException("The Organization with id given doesn't exist in the base.");
+        } else
+            equipmentRepository.deleteById(id);
     }
 }
